reverted back to using smaller integers

+++ embedded/include/code_extraction.h
@@ -17,7 +17,7 @@ void img_bit_matrix_conversion(BitMatrix32 dst, BitMatrix32 mask, const ImageMat
         PIXEL_TYPE low_thresh, PIXEL_TYPE high_thresh);
 
 void bm32_extract_axis_codes(AxisCode* row_code, AxisCode* col_code, BitMatrix32 matrix,
-        BitMatrix32 mask, uint32_t min_samples);
+        BitMatrix32 mask, uint8_t min_samples);
 
 AxisCode bm32_extract_column_code(uint32_t initial_row_guess, const BitMatrix32 matrix,
-        const BitMatrix32 mask, uint32_t min_row_samples);
+        const BitMatrix32 mask, uint8_t min_row_samples);
