+++ embedded/src/utils/image_utils_extra.c
@@ -19,11 +19,11 @@ void img_hough_line_transform(ImageMatrix dst, const ImageMatrix src) {
     float angle_resolution = M_PI * 0.5f / dst.n_rows;
     float scale_to_index =
             dst.n_cols / sqrtf((src.n_rows * src.n_rows) + (src.n_cols * src.n_cols));
-    for (int32_t i = 0; i < dst.n_rows; ++i) {
+    for (uint16_t i = 0; i < dst.n_rows; ++i) {
         float sin = sinf(i * angle_resolution);
         float cos = cosf(i * angle_resolution);
         FOR_EACH_PIXEL(src, ) {
-            PIXEL(dst, i, (int32_t)(((sin * row) + (cos * col)) * scale_to_index)) +=
+            PIXEL(dst, i, (int16_t)(((sin * row) + (cos * col)) * scale_to_index)) +=
                     PIXEL(src, row, col);
         }
     }
